YAY FINALLY some progress
