+++ VAST/VAST.vcxproj.filters
@@ -18,9 +18,6 @@
     <ClCompile Include="cpp\AV.cpp">
       <Filter>Resource Files</Filter>
     </ClCompile>
-    <ClCompile Include="cpp\Environment.cpp">
-      <Filter>Resource Files</Filter>
-    </ClCompile>
     <ClCompile Include="cpp\main.cpp">
       <Filter>Resource Files</Filter>
     </ClCompile>
@@ -56,16 +53,7 @@
     <ClInclude Include="h\VType.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="h\VView.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="h\Environment.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
     <ClInclude Include="h\Obstacle.h">
-      <Filter>Header Files</Filter>
-    </ClInclude>
-    <ClInclude Include="h\vector3.h">
       <Filter>Header Files</Filter>
     </ClInclude>
     <ClInclude Include="h\socket.h">
