--- VAST/cpp/Environment.cpp
@@ -1,30 +1,0 @@
-#include "..\h\Environment.h"
-#include <iostream>
-
-
-//Adding SUMO Socket connection stuff
-
-Environment::Environment(vector3 _dimensions, int _numObstacles)
-{
-
-}
-
-Environment::~Environment()
-{
-
-}
-void Environment::addObstacle()
-{
-
-}
-/*void Environment::updateAV(Command update)
-{
-
-}*/
-void Environment::update()
-{
-
-}
-
-
-
