+++ VASTTesting/Environment.cpp
@@ -1,0 +1,27 @@
+#include "Environment.h"
+
+
+//Adding SUMO Socket connection stuff
+
+Environment::Environment()
+{
+
+}
+
+Environment::~Environment()
+{
+
+}
+
+void Environment::addObstacle()
+{
+
+}
+
+void Environment::update(timestamp t, dataMap* dataMap)
+{
+
+}
+
+
+
